requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.3
MarkupSafe==3.0.2
redis==5.0.3
requests==2.32.0
urllib3==2.3.0
Werkzeug==3.1.3
